requirements.txt
pydantic==1.9.0
requests==2.27.1
typing_extensions==4.0.1
pyyaml==6.0